extras: Suppress build warnings
On x86_64, when printing a __u64 using PRIu64, gcc throws a warning:
warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’.

This patch silences the build warning by casting the argument as uint64_t.

+++ extras/find_allocation_fragments.c
@@ -89,7 +89,8 @@ static int print_group(struct ocfs2_group_desc *gd)
 
 		free = end - start;
 
-		printf("%-6u   %-6u   %"PRIu64"\n", free, start, gd->bg_blkno);
+		printf("%-6u   %-6u   %"PRIu64"\n", free, start,
+		       (uint64_t)gd->bg_blkno);
 
 		if (free < FREE_BIT_STATS)
 			free_bit_stats[free]++;
